requirements.txt
psd-tools
Pillow
python-telegram-bot==13.15
python-dotenv
